requirements.txt
fastapi
uvicorn
psycopg2-binary
openai
gitpython
python-dotenv
pgvector